Move chain transactions API out of Node into its own class.
Still avoiding modifying tests for the time being.
