+++ tests/TestHarness/CMakeLists.txt
@@ -5,3 +5,4 @@ configure_file(Node.py Node.py COPYONLY)
 configure_file(Cluster.py Cluster.py COPYONLY)
 configure_file(TestHelper.py TestHelper.py COPYONLY)
 configure_file(queries.py queries.py COPYONLY)
+configure_file(transactions.py transactions.py COPYONLY)
