+++ tests/TestHarness/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ['Node', 'Cluster', 'WalletMgr', 'testUtils', 'TestHelper', 'queries']
+__all__ = ['Node', 'Cluster', 'WalletMgr', 'testUtils', 'TestHelper', 'queries', 'transactions']
 
 from .Cluster import Cluster
 from .Node import Node
